requirements.txt.txt
tensorflow==2.4.1
streamlit==0.76.0
opencv-contrib-python-headless==4.4.0.46
numpy==1.19.2
Pillow==8.1.0
scikit-image==0.17.2
Scikit-learn==0.24.1
matplotlib==3.3.2
bing_image_downloader==1.0.4